feat: 添加全局注册request功能
